yolo5/requirements.txt
flask
pyyaml
loguru
pymongo
torch
torchvision
torchaudio
numpy
pillow
ultralytics
requests
boto3